requirements.txt

# python -m pip install --index-url=https://blpapi.bloomberg.com/repository/releases/python/simple/ blpapi
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
pysabr>=0.3.0
py_vollib>=1.0.1
python-dateutil>=2.8.0
holidays>=0.35
python-dotenv>=1.0.0
google-adk>=0.1.0
